requirements.txt
annotated-types==0.7.0
anyio==3.7.1
backports-asyncio-runner==1.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.7
pydantic==2.4.2
pydantic-core==2.10.1
pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.0.0
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
tomli==2.2.1
typing-extensions==4.14.1
urllib3==2.5.0
uvicorn==0.23.2